refactor(gui): delete some not need args, shared_config use global var, and not export this, for do this, not set run arg(storage_secret)
