+++ gui/define.py
@@ -1,0 +1,5 @@
+from modules.configs.MyConfig import MyConfigger
+
+
+# 构造一个config，用于在tab间共享softwareconfigdict
+gui_shared_config = MyConfigger()
