+++ gui/pages/Setting_task_order.py
@@ -25,9 +25,8 @@ def set_task_order(config, real_taskname_to_show_taskname):
             with ui.row():
                 ui.label(f'{config.get_text("config_task")} {i+1}:')
                 atask = ui.select(real_taskname_to_show_taskname,
-                            value=config.userconfigdict["TASK_ORDER"][i],
-                            on_change=lambda v,i=i: config.userconfigdict["TASK_ORDER"].__setitem__(i, v.value),
-                            )
+                                  value=config.userconfigdict["TASK_ORDER"][i],
+                                  on_change=lambda v: config.userconfigdict["TASK_ORDER"].__setitem__(i, v.value))
                 acheck = ui.checkbox(config.get_text("button_enable"), value=config.userconfigdict["TASK_ACTIVATE"][i], on_change=lambda v,i=i: config.userconfigdict["TASK_ACTIVATE"].__setitem__(i, v.value))
                 if i==0:
                     atask.set_enabled(False)
